[change] - separete constructor, cover test
